Rename cmake_configure_ci.py -> cmake_ci.py.
* Also turns it into a generic wrapper around cmake. Will be useful for supporting windows.
